web_scraper_and_flask_app/api/requirements.txt
click==8.1.3
Flask==2.2.2
flask-log==0.1.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.0.post0
gunicorn==20.1.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
python-dotenv==0.21.0
SQLAlchemy==1.4.42
Werkzeug==2.2.2
zipp==3.10.0
